refactor: разнести проверки по слоям.

+++ src/main/java/ru/yandex/practicum/filmorate/storage/impl/ReviewDbStorage.java
@@ -6,6 +6,7 @@ import org.springframework.jdbc.core.JdbcOperations;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
 import org.springframework.stereotype.Repository;
+import ru.yandex.practicum.filmorate.exeptions.NotFoundException;
 import ru.yandex.practicum.filmorate.exeptions.ValidationException;
 import ru.yandex.practicum.filmorate.model.Review;
 import ru.yandex.practicum.filmorate.storage.ReviewStorage;
@@ -22,12 +23,16 @@ public class ReviewDbStorage implements ReviewStorage {
 
     private final JdbcOperations jdbcOperations;
     private final ReviewRowMapper reviewRowMapper = new ReviewRowMapper();
+    private final UserDbStorage userDbStorage;
+    private final FilmDbStorage filmDbStorage;
 
     /**
      * ALG_1
      */
     @Override
     public Review create(Review review) {
+        userDbStorage.checkContainsUser(review.getUserId());
+        filmDbStorage.checkContainsFilm(review.getFilmId());
         if (review.getReviewId() != null) throw new ValidationException("Поле id у отзыва не пустое");
         SimpleJdbcInsert simpleJdbcInsert = new SimpleJdbcInsert((JdbcTemplate) jdbcOperations);
         simpleJdbcInsert.withTableName("reviews").usingGeneratedKeyColumns("review_id");
@@ -46,12 +51,16 @@ public class ReviewDbStorage implements ReviewStorage {
      */
     @Override
     public Review update(Review review) {
+        userDbStorage.checkContainsUser(review.getUserId());
+        filmDbStorage.checkContainsFilm(review.getFilmId());
         Review updatedReview = getById(review.getReviewId());
         String sql = "UPDATE reviews SET content = ?, is_positive = ? WHERE review_id = ?";
-        jdbcOperations.update(sql, review.getContent(), review.getIsPositive(), review.getReviewId());
+        int rows = jdbcOperations.update(sql, review.getContent(), review.getIsPositive(), review.getReviewId());
         updatedReview.setContent(review.getContent());
         updatedReview.setIsPositive(review.getIsPositive());
-
+        if (rows == 0) {
+            throw new NotFoundException("Invalid Review ID:  " + review.getReviewId());
+        }
         return updatedReview;
     }
 
@@ -61,8 +70,11 @@ public class ReviewDbStorage implements ReviewStorage {
     @Override
     public Review delete(Long reviewId) {
         Review review = getById(reviewId);
+        int rows = jdbcOperations.update("DELETE FROM reviews WHERE review_id = ?", reviewId);
+        if (rows == 0) {
+            throw new NotFoundException("Invalid Review ID:  " + review.getReviewId());
+        }
         jdbcOperations.update("DELETE FROM reviews_like WHERE review_id = ?", reviewId);
-        jdbcOperations.update("DELETE FROM reviews WHERE review_id = ?", reviewId);
         return review;
     }
 
@@ -77,7 +89,11 @@ public class ReviewDbStorage implements ReviewStorage {
                 "LEFT JOIN reviews_like rl ON r.review_id = rl.review_id " +
                 "WHERE r.review_id = ? " +
                 "GROUP BY r.review_id";
-        return jdbcOperations.queryForObject(sql, reviewRowMapper, reviewId);
+        return jdbcOperations
+                .query(sql, reviewRowMapper, reviewId)
+                .stream()
+                .findFirst()
+                .orElseThrow(() -> new NotFoundException("Review with id = " + reviewId + " not found"));
     }
 
     /**
@@ -98,8 +114,12 @@ public class ReviewDbStorage implements ReviewStorage {
      */
     @Override
     public void addLike(Long reviewId, Long userId) {
+        userDbStorage.checkContainsUser(userId);
         String sql = "INSERT INTO reviews_like (review_id, user_id, is_useful) VALUES(?, ?, true)";
-        jdbcOperations.update(sql, reviewId, userId);
+        int rows = jdbcOperations.update(sql, reviewId, userId);
+        if (rows == 0) {
+            throw new NotFoundException("Invalid Review ID:  " + reviewId);
+        }
     }
 
     /**
@@ -107,8 +127,12 @@ public class ReviewDbStorage implements ReviewStorage {
      */
     @Override
     public void addDislike(Long reviewId, Long userId) {
+        userDbStorage.checkContainsUser(userId);
         String sql = "INSERT INTO reviews_like (review_id, user_id, is_useful) VALUES(?, ?, false)";
-        jdbcOperations.update(sql, reviewId, userId);
+        int rows = jdbcOperations.update(sql, reviewId, userId);
+        if (rows == 0) {
+            throw new NotFoundException("Invalid Review ID:  " + reviewId);
+        }
     }
 
     /**
@@ -116,8 +140,12 @@ public class ReviewDbStorage implements ReviewStorage {
      */
     @Override
     public void deleteLike(Long reviewId, Long userId) {
+        userDbStorage.checkContainsUser(userId);
         String sql = "DELETE FROM reviews_like WHERE review_id = ? AND user_id = ? AND is_useful = true";
-        jdbcOperations.update(sql, reviewId, userId);
+        int rows = jdbcOperations.update(sql, reviewId, userId);
+        if (rows == 0) {
+            throw new NotFoundException("Invalid Review ID:  " + reviewId);
+        }
     }
 
     /**
@@ -125,7 +153,11 @@ public class ReviewDbStorage implements ReviewStorage {
      */
     @Override
     public void deleteDislike(Long reviewId, Long userId) {
+        userDbStorage.checkContainsUser(userId);
         String sql = "DELETE FROM reviews_like WHERE review_id = ? AND user_id = ? AND is_useful = false";
-        jdbcOperations.update(sql, reviewId, userId);
+        int rows = jdbcOperations.update(sql, reviewId, userId);
+        if (rows == 0) {
+            throw new NotFoundException("Invalid Review ID:  " + reviewId);
+        }
     }
 }
